LayoutTests/media/controls/airplay-picker.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../media-file.js"></script>
        <script src="controls-test-helpers.js"></script>
        <script>
            var tester = new ControlsTest()
                .whenReady(runTests)
                .start();

            function runTests()
            {
                var currentState = tester.currentState;
                tester.test("We are using the apple idiom")
                    .value(currentState.idiom)
                    .isEqualTo("apple");

                tester.startNewSection("Get the wireless device picker button");
                var statusState = tester.stateForControlsElement("AppleTV Device Picker");

                tester.test("Name is 'AppleTV Device Picker'")
                    .value(statusState.name)
                    .isEqualTo("AppleTV Device Picker");

                tester.test("Has hidden class")
                    .value(statusState.className)
                    .isEqualTo("hidden");

                tester.test("CSS display: none")
                    .value(statusState.computedStyle.display)
                    .isEqualTo("none");

                tester.test("Aria label is 'Choose Wireless Display'")
                    .value(statusState.ariaLabel)
                    .isEqualTo("Choose Wireless Display");

                tester.end();
            }

        </script>
    </head>
    <body>
        <p>This tests for the wireless device picker icon.</p>
        <p>This test only runs in DRT!</p>
        <video controls></video>
    </body>
</html>
